# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SafeSend marketplace escrow API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deals/{dealId}/tracking:
    get:
      operationId: getTracking
      tags: [tracking]
      summary: Get shipment tracking events for a deal
      parameters:
        - name: dealId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracking events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingResult"
  /deals/{dealId}/tracking-public:
    get:
      operationId: getPublicTracking
      tags: [tracking]
      summary: Public shipment tracking — no auth required, returns safe subset only
      parameters:
        - name: dealId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTrackingResult"
        "404":
          description: Deal not found
...
  /admin/deals:
    get:
      operationId: adminListDeals
      tags: [admin]
      summary: Admin view of all deals with filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: disputed
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated deals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDealsResult"
  /admin/deals/{dealId}/resolve-dispute:
    post:
      operationId: resolveDispute
      tags: [admin]
      summary: Admin resolves a dispute (refund buyer or release to seller)
      parameters:
        - name: dealId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDisputeBody"
      responses:
        "200":
          description: Dispute resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
...
  /auth/request-link:
    post:
      operationId: requestLoginLink
      tags: [auth]
      summary: Send a magic sign-in link to an email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestLoginLinkBody"
      responses:
        "200":
          description: Link sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestLoginLinkResponse"
        "400":
          description: Invalid email
        "500":
          description: Email delivery failed
  /auth/verify:
    post:
      operationId: verifyLoginLink
      tags: [auth]
      summary: Verify a magic-link token and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTokenBody"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "400":
          description: Invalid or expired token
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Return the currently authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    DealState:
      type: string
      enum:
        - pending_seller_acceptance
        - pending_buyer_confirmation
        - created
        - funded
        - shipped
        - delivered
        - complete
        - disputed
        - cancelled
        - refunded
    Deal:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        amountNzd:
          type: number
          description: Item price in NZD (excluding fee)
        feeNzd:
          type: number
          description: Platform fee in NZD (4% with $5 min)
        kycFeeNzd:
          type: number
          description: Identity verification fee in NZD ($2.50 for deals >= $500 NZD, else 0)
        totalNzd:
          type: number
          description: Total charged to buyer (amountNzd + feeNzd + kycFeeNzd)
        state:
          $ref: "#/components/schemas/DealState"
        buyerEmail:
          type: string
        sellerEmail:
          type: string
        trackingNumber:
          type: string
          nullable: true
        courierSlug:
          type: string
          nullable: true
        disputeReason:
          type: string
          nullable: true
        stripePaymentIntentId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        fundedAt:
          type: string
          format: date-time
          nullable: true
        shippedAt:
          type: string
          format: date-time
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
          description: Optimistic concurrency version
        signatureRequired:
          type: boolean
          description: Whether a courier signature is required at delivery
        stripeTransferId:
          type: string
          nullable: true
          description: Stripe transfer ID when funds were released to seller (for reconciliation)
        stripeRefundId:
          type: string
          nullable: true
          description: Stripe refund ID when a refund was issued to buyer (for reconciliation)
        stripeTransferError:
          type: string
          nullable: true
          description: Last Stripe transfer error message, if any
        shipmentVerificationStatus:
          type: string
          enum: [pending, verified, flagged]
          description: TrackingMore courier scan status after mark-shipped
        buyerPhone:
          type: string
          nullable: true
          description: Buyer mobile for SMS alerts only — never exposed to seller
        sellerPhone:
          type: string
          nullable: true
          description: Seller mobile for SMS alerts only — never exposed to buyer
        itemUrl:
          type: string
          nullable: true
          description: Original marketplace listing URL for product identification
        referenceNumber:
          type: string
          nullable: true
          description: Purchase order or external reference number
        invoiceNumber:
          type: string
          nullable: true
          description: "Auto-generated contract/invoice number for dispute callbacks. Format: SS-XXXXXXX"
        creatorRole:
          type: string
          enum: [buyer, seller]
          nullable: true
          description: Whether the deal was created by the buyer or the seller
      required:
        - id
        - title
        - description
        - amountNzd
        - feeNzd
        - kycFeeNzd
        - totalNzd
        - state
        - buyerEmail
        - sellerEmail
        - createdAt
        - updatedAt
...
    TrackingEvent:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        location:
          type: string
          nullable: true
        status:
          type: string
      required:
        - timestamp
        - message
        - status
    TrackingResult:
      type: object
      properties:
        trackingNumber:
          type: string
          nullable: true
        courierSlug:
          type: string
          nullable: true
        currentStatus:
          type: string
          nullable: true
        events:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
      required:
        - events
    PublicTrackingResult:
      type: object
      description: Safe public subset of deal + tracking data — no financial or contact info
      properties:
        id:
          type: string
        title:
          type: string
        state:
          $ref: "#/components/schemas/DealState"
        trackingNumber:
          type: string
          nullable: true
        courierSlug:
          type: string
          nullable: true
        signatureRequired:
          type: boolean
        shipmentVerificationStatus:
          type: string
          enum: [pending, verified, flagged]
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        shippedAt:
          type: string
          format: date-time
          nullable: true
        buyerEmailMasked:
          type: string
          nullable: true
          description: Partially masked buyer email for isBuyer check (e.g. j***@example.com)
        referenceNumber:
          type: string
          nullable: true
        itemUrl:
          type: string
          nullable: true
        currentStatus:
          type: string
          nullable: true
        events:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
      required:
        - id
        - title
        - state
        - signatureRequired
        - shipmentVerificationStatus
        - events
...
    AdminDealsResult:
      type: object
      properties:
        deals:
          type: array
          items:
            $ref: "#/components/schemas/Deal"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - deals
        - total
        - page
        - limit
    ResolveDisputeBody:
      type: object
      properties:
        resolution:
          type: string
          enum: [refund_buyer, release_to_seller]
        adminNote:
          type: string
      required:
        - resolution
...
    RequestLoginLinkBody:
      type: object
      properties:
        email:
          type: string
          format: email
        returnPath:
          type: string
      required:
        - email
    VerifyTokenBody:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    AuthMeResponse:
      type: object
      properties:
        email:
          type: string
          format: email
        verifiedAt:
          type: string
          format: date-time
          description: When this email was first verified via magic link on SafeSend
      required:
        - email
        - verifiedAt
    RequestLoginLinkResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        devLink:
          type: string
          description: Only present in development when RESEND_API_KEY is not set. The magic link URL for testing without an email client.
      required:
        - message
...